# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": " Api documentation",
        "description": "Documentation avec Swagger",
        "version": "1.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Register"
                ],
                "summary": "Register EndPoint",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "prenom",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "elgarrai"
                                    },
                                    "prenom": {
                                        "type": "string",
                                        "example": "Hamza"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "Hamza@gmail.com"
                                    },
                                    "pasword": {
                                        "type": "string",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "user registred"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Login EndPoint",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "token"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "hamza@emple.com"
                                    },
                                    "password": {
                                        "type": "integer",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "user registred"
                    }
                }
            }
        },
        "/api/password/forget": {
            "post": {
                "tags": [
                    "Forget"
                ],
                "summary": "Forget EndPoint",
                "operationId": "686e313af2e7a9413c06bf0adfa44b6a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "hamza@emple.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "a confirmation code is sent to user"
                    }
                }
            }
        },
        "/api/password/reset": {
            "post": {
                "tags": [
                    "Reset"
                ],
                "summary": "Reset EndPoint",
                "operationId": "5f9daa4e2d9a09acda465d6b8455a1cd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "token",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "hamza@emple.com"
                                    },
                                    "token": {
                                        "type": "string",
                                        "example": "hamza@emple.com"
                                    },
                                    "password": {
                                        "type": "integer",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "password reset succesfuly"
                    }
                }
            }
        },
...
    },
...
}